fix: split oversized test_test_filter.py — move step7 tests to test_test_filter_step7.py
test_test_filter.py exceeded the 1000-line limit (1074 lines) enforced by
test_test_suite_oversized_files_split. The seven step 7 tests (S11–S17) are
extracted into a standalone test_test_filter_step7.py (212 lines), bringing
test_test_filter.py to 874 lines.
